Constant-time stepping not so constant.
Good advantage on a single motion with thousands of phrases.
Barely better on motions with ~10 phrases (assuming we use double for Time in both cases).
Requires double for Time type to be accurate.
Slower when doing many motions with 2 phrases.
Probably the forward-only memory traversal of the naive Sequence approach is cache-friendlier.