Add Timeline cleared callback.
Can be used to destroy the timeline once all of its animations finish.
For example, if you have an object that manages transitions on its
own timeline, you can destroy that object with its timeline's setClearedFn,
allowing the object to manage its own exit transition on that timeline.

Of course, you could avoid the destruction problem by using a timeline that
outlives the object. In that case, you need a bit more communication between
levels as to where the timeline lives.

Will think about this more, but for now the cleared callback will exist.