Explicitly define move ctors for Windows.
We need to define move constructors for types holding Outputs by name
because VS2013 isn't compliant with the C++11 spec.
http://msdn.microsoft.com/en-us/library/hh567368.aspx

Clang does the right thing and generates default move constructors for
us.

In the worst case, we should be able to explicitly declare the move
constructor as `default`, but VS2013 doesn't support this (yet?).
http://msdn.microsoft.com/en-us/library/dn457344.aspx